requirements.txt
sympy
networkx
matplotlib
pygraphviz
Pillow